修改文件路径，匹配linux系统，现在的文件上传是不支持windows的，以后会调整

+++ src/com/bkweb/modules/menu/service/MenuFindService.java
@@ -28,9 +28,11 @@ public class MenuFindService extends CrudService<MenuFindDao, MenuFind> {
 				FileUtils.deleteFile(file.getAbsolutePath());
 				return false;
 			}
-			String url = dao.get(menuFind).getLocalUrl();
-			if (!StringUtils.isEmpty(url)) {
-				FileUtils.deleteFile(url);
+			if (!StringUtils.isEmpty(menuFind.getId())) {
+				String url = dao.get(menuFind).getLocalUrl();
+				if (!StringUtils.isEmpty(url)) {
+					FileUtils.deleteFile(url);
+				}
 			}
 			menuFind.setImage(FileUploadUtils.getDefaultImgUrl(request, file.getName()));
 			menuFind.setLocalUrl(file.getAbsolutePath());
